# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Minecraft Nexus Panel (MCNP) API
  version: 0.1.0-draft
  summary: Panel API for Web, Desktop, and Mobile clients
  description: |
    Draft v1 contract. Browser clients use a secure session cookie plus CSRF
    token for writes. Native clients use bearer access tokens and rotating
    refresh tokens.
servers:
  - url: /api/v1
    description: Current Panel
...
security:
  - cookieAuth: [ ]
  - bearerAuth: [ ]
paths:
...
  /cores/{coreId}/instances/{proxyInstanceId}/proxy-subservers:
    parameters:
      - $ref: '#/components/parameters/CoreId'
      - $ref: '#/components/parameters/ProxyInstanceId'
    get:
      tags: [ Proxy ]
      operationId: listProxySubservers
      summary: List the backend instances configured for a proxy
      responses:
        '200':
          description: Proxy backend relationships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxySubserverPage'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      tags: [ Proxy ]
      operationId: upsertProxySubserver
      summary: Create or replace a proxy backend relationship
      security:
        - cookieAuth: [ ]
          csrfToken: [ ]
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxySubserver'
      responses:
        '200':
          description: Proxy backend relationship saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxySubserver'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionRequired'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /cores/{coreId}/instances/{proxyInstanceId}/proxy-subservers/{subserverId}:
    parameters:
      - $ref: '#/components/parameters/CoreId'
      - $ref: '#/components/parameters/ProxyInstanceId'
      - $ref: '#/components/parameters/SubserverId'
    delete:
      tags: [ Proxy ]
      operationId: deleteProxySubserver
      summary: Remove a proxy backend relationship
      security:
        - cookieAuth: [ ]
          csrfToken: [ ]
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Proxy backend relationship removed
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionRequired'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mcnp_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  parameters:
...
    CoreId:
      name: coreId
      in: path
      required: true
      schema:
        type: string
...
    ProxyInstanceId:
      name: proxyInstanceId
      in: path
      required: true
      schema:
        type: string
    SubserverId:
      name: subserverId
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource does not exist or is not visible
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource or state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    PreconditionRequired:
      description: Required idempotency precondition is missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServiceUnavailable:
      description: Panel or Core is temporarily unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ProxySubserver:
      type: object
      required: [ id, name, targetInstanceId, host, port, enabled ]
      properties:
        id:
          type: string
          pattern: '^[a-zA-Z0-9][a-zA-Z0-9_-]{0,63}$'
        name:
          type: string
          minLength: 1
          maxLength: 128
        targetInstanceId:
          type: string
        host:
          type: string
          minLength: 1
          maxLength: 253
        port:
          type: integer
          minimum: 1
          maximum: 65535
        enabled:
          type: boolean
      additionalProperties: false
    ProxySubserverPage:
      type: object
      required: [ items ]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProxySubserver'
...
    Error:
      type: object
      required: [ code, message, requestId, retryable ]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [ error ]
      properties:
        error:
          $ref: '#/components/schemas/Error'